refactor away posteditview
