+++ Apps/iOS/Postalgic/Views/Post List/EmbedView.swift
@@ -196,7 +196,7 @@ struct WebViewContainer: UIViewRepresentable {
     VStack(spacing: 20) {
         // YouTube Embed
         EmbedView(embed: Embed(
-            url: "https://www.youtube.com/watch?v=dQw4w9WgXcQ",
+            url: "https://www.youtube.com/watch?v=RoSQqtgCZss",
             type: .youtube,
             position: .above
         ))
