+++ Apps/iOS/Postalgic/Views/Post List/PostFormView.swift
@@ -2,7 +2,7 @@
 //  PostFormView.swift
 //  Postalgic
 //
-//  Created by Brad Root on 4/19/25.
+//  Created by Brad Root on 4/23/25.
 //
 
 import SwiftData
@@ -13,32 +13,41 @@ struct PostFormView: View {
     @Environment(\.dismiss) private var dismiss
     @Query private var allTags: [Tag]
     @Query private var allCategories: [Category]
-
-    var blog: Blog
-
+    
+    // Either blog (for new post) or post (for editing) will be set
+    var blog: Blog?
+    var post: Post?
+    var isEditing: Bool
+    
+    private var currentBlog: Blog? {
+        return isEditing ? post?.blog : blog
+    }
+    
     private var blogTags: [Tag] {
-        return allTags.filter { $0.blog?.id == blog.id }
+        guard let blogId = currentBlog?.id else { return [] }
+        return allTags.filter { $0.blog?.id == blogId }
     }
-
+    
     private var blogCategories: [Category] {
-        return allCategories.filter { $0.blog?.id == blog.id }
+        guard let blogId = currentBlog?.id else { return [] }
+        return allCategories.filter { $0.blog?.id == blogId }
     }
-
-    @State private var title = ""
-    @State private var content = ""
+    
+    @State private var title: String
+    @State private var content: String
     @State private var tagInput = ""
     @State private var selectedTags: [Tag] = []
     @State private var selectedCategory: Category?
-    @State private var isDraft = false
+    @State private var isDraft: Bool
     @State private var showingCategoryManagement = false
     @State private var showingSuggestions = false
     @State private var showingEmbedForm = false
     @State private var newPost: Post? = nil
-
+    
     private var existingTagNames: [String] {
         return blogTags.map { $0.name }
     }
-
+    
     private var filteredTags: [Tag] {
         if tagInput.isEmpty {
             return blogTags.sorted {
@@ -52,21 +61,45 @@ struct PostFormView: View {
             .sorted { $0.name.lowercased() < $1.name.lowercased() }
         }
     }
-
+    
+    // Initialize for new post
+    init(blog: Blog) {
+        self.blog = blog
+        self.post = nil
+        self.isEditing = false
+        
+        _title = State(initialValue: "")
+        _content = State(initialValue: "")
+        _isDraft = State(initialValue: false)
+    }
+    
+    // Initialize for editing post
+    init(post: Post) {
+        self.blog = nil
+        self.post = post
+        self.isEditing = true
+        
+        _title = State(initialValue: post.title ?? "")
+        _content = State(initialValue: post.content)
+        _isDraft = State(initialValue: post.isDraft)
+        _selectedTags = State(initialValue: post.tags)
+        _selectedCategory = State(initialValue: post.category)
+    }
+    
     var body: some View {
         NavigationStack {
             Form {
                 Section("Post Details") {
                     TextField("Title (optional)", text: $title)
                     Toggle("Save as Draft", isOn: $isDraft)
-
+                    
                     HStack {
                         Picker("Category", selection: $selectedCategory) {
                             Text("None").tag(Category?.none)
-
+                            
                             if !blogCategories.isEmpty {
                                 Divider()
-
+                                
                                 ForEach(
                                     blogCategories.sorted { $0.name < $1.name }
                                 ) { category in
@@ -75,7 +108,7 @@ struct PostFormView: View {
                             }
                         }
                         .frame(maxWidth: .infinity, alignment: .leading)
-
+                        
                         Button(action: {
                             showingCategoryManagement = true
                         }) {
@@ -83,14 +116,14 @@ struct PostFormView: View {
                         }
                     }
                 }
-
+                
                 Section("Content") {
                     TextEditor(text: $content)
                         .frame(minHeight: 200)
                 }
                 
                 Section("Embed") {
-                    if let post = newPost, let embed = post.embed {
+                    if let postToUse = isEditing ? post : newPost, let embed = postToUse.embed {
                         VStack(alignment: .leading, spacing: 12) {
                             // Embed info row
                             HStack {
@@ -126,9 +159,9 @@ struct PostFormView: View {
                             .tint(.blue)
                             
                             Button(action: {
-                                if let embed = post.embed {
+                                if let embed = postToUse.embed {
                                     modelContext.delete(embed)
-                                    post.embed = nil
+                                    postToUse.embed = nil
                                 }
                             }) {
                                 HStack {
@@ -142,8 +175,8 @@ struct PostFormView: View {
                         }
                     } else {
                         Button(action: {
-                            // Create a temporary post if needed
-                            if newPost == nil {
+                            // Create a temporary post if needed for new posts
+                            if !isEditing && newPost == nil {
                                 newPost = Post(
                                     title: title.isEmpty ? nil : title,
                                     content: content,
@@ -160,7 +193,7 @@ struct PostFormView: View {
                         .tint(.blue)
                     }
                 }
-
+                
                 Section("Tags") {
                     HStack {
                         TextField("Add tags...", text: $tagInput)
@@ -176,13 +209,13 @@ struct PostFormView: View {
                                 }
                                 showingSuggestions = true
                             }
-
+                        
                         Button(action: addTag) {
                             Image(systemName: "plus.circle.fill")
                         }
                         .disabled(tagInput.isEmpty)
                     }
-
+                    
                     if showingSuggestions && !filteredTags.isEmpty {
                         ScrollView(.horizontal, showsIndicators: false) {
                             HStack {
@@ -210,7 +243,7 @@ struct PostFormView: View {
                             .padding(.vertical, 5)
                         }
                     }
-
+                    
                     if !selectedTags.isEmpty {
                         ScrollView(.horizontal, showsIndicators: false) {
                             HStack {
@@ -239,7 +272,7 @@ struct PostFormView: View {
                     }
                 }
             }
-            .navigationTitle("Create New Post")
+            .navigationTitle(isEditing ? "Edit Post" : "Create New Post")
             .toolbar {
                 ToolbarItem(placement: .cancellationAction) {
                     Button("Cancel") {
@@ -248,27 +281,35 @@ struct PostFormView: View {
                 }
                 ToolbarItem(placement: .confirmationAction) {
                     Button("Save") {
-                        addPost()
+                        if isEditing {
+                            updatePost()
+                        } else {
+                            addPost()
+                        }
                         dismiss()
                     }
                     .disabled(content.isEmpty)
                 }
             }
             .sheet(isPresented: $showingCategoryManagement) {
-                CategoryManagementView(blog: blog)
+                if let blog = currentBlog {
+                    CategoryManagementView(blog: blog)
+                }
             }
             .sheet(isPresented: $showingEmbedForm) {
-                if let post = newPost {
+                if isEditing, let post = post {
+                    EmbedFormView(post: post)
+                } else if let post = newPost {
                     EmbedFormView(post: post)
                 }
             }
         }
     }
-
+    
     private func addTag() {
         let trimmed = tagInput.trimmingCharacters(in: .whitespacesAndNewlines)
             .lowercased()
-        if !trimmed.isEmpty {
+        if !trimmed.isEmpty, let blog = currentBlog {
             // Check if tag already exists for this blog (case insensitive)
             if let existingTag = blogTags.first(where: {
                 $0.name.lowercased() == trimmed
@@ -287,8 +328,54 @@ struct PostFormView: View {
             tagInput = ""
         }
     }
-
+    
+    private func updatePost() {
+        guard let post = post else { return }
+        
+        // Update post properties
+        post.title = title.isEmpty ? nil : title
+        post.content = content
+        post.isDraft = isDraft
+        
+        // Handle category changes
+        if post.category != selectedCategory {
+            // Remove post from previous category
+            if let oldCategory = post.category {
+                if let index = oldCategory.posts.firstIndex(where: {
+                    $0.id == post.id
+                }) {
+                    oldCategory.posts.remove(at: index)
+                }
+            }
+            
+            // Add post to new category
+            post.category = selectedCategory
+            if let newCategory = selectedCategory {
+                newCategory.posts.append(post)
+            }
+        }
+        
+        // Handle tag changes
+        // First, remove all existing tag relationships
+        for tag in post.tags {
+            if let index = tag.posts.firstIndex(where: { $0.id == post.id }) {
+                tag.posts.remove(at: index)
+            }
+        }
+        
+        // Clear post's tags array
+        post.tags.removeAll()
+        
+        // Now add all selected tags
+        for tag in selectedTags {
+            post.tags.append(tag)
+            tag.posts.append(post)
+        }
+    }
+    
     private func addPost() {
+        guard let blog = blog else { return }
+        
         var postToSave: Post
         
         if let existingPost = newPost {
@@ -306,19 +393,19 @@ struct PostFormView: View {
             )
             modelContext.insert(postToSave)
         }
-
+        
         // Add category to post if selected
         if let category = selectedCategory {
             postToSave.category = category
             category.posts.append(postToSave)
         }
-
+        
         // Add tags to post
         for tag in selectedTags {
             postToSave.tags.append(tag)
             tag.posts.append(postToSave)
         }
-
+        
         // Add to blog
         blog.posts.append(postToSave)
     }
@@ -327,7 +414,19 @@ struct PostFormView: View {
 #Preview {
     PostFormView(blog: Blog(name: "Test Blog", url: "https://example.com"))
         .modelContainer(
-            for: [Blog.self, Post.self, Tag.self, Category.self],
+            for: [Blog.self, Post.self, Tag.self, Category.self, Embed.self],
+            inMemory: true
+        )
+}
+
+#Preview {
+    let testPost = Post(
+        title: "Test Post",
+        content: "This is a test post with **bold** and *italic* text."
+    )
+    return PostFormView(post: testPost)
+        .modelContainer(
+            for: [Post.self, Tag.self, Category.self, Blog.self, Embed.self],
             inMemory: true
         )
 }
